Change to select random gif, add chipmunk stare
